+++ resources/assets/js/app.js
@@ -1,7 +1,13 @@
 import Vue from 'vue'
+import PortalVue from 'portal-vue'
+import axios from 'axios'
 import './components'
 
 window.$ = window.jQuery = require('jquery')
+
+axios.defaults.headers['X-CSRF-TOKEN'] = document.querySelector('meta[name="csrf-token"]').getAttribute('content')
+
+Vue.use(PortalVue)
 
 $(document).ready(function () {
     const searchForm = $('#search-form')