+++ resources/assets/js/components/index.js
@@ -1,4 +1,8 @@
 import Vue from 'vue'
 import FormSelect from './FormSelect'
+import MediaManager from './MediaManager/MediaManager'
+import FileInput from './FileInput'
 
 Vue.component('form-select', FormSelect)
+Vue.component('media-manager', MediaManager)
+Vue.component('file-input', FileInput)